refactor: :truck: component MenuMobile moved to src/components/containers/header
